seperate out xaddress and address, move checksum test to AddressCodecTest
